Increase speed of *.plugin.zsh file

+++ zsh-navigation-tools.plugin.zsh
@@ -23,13 +23,21 @@ if ! test -d "$CONFIG_DIR"; then
     mkdir "$CONFIG_DIR"
 fi
 
+# 9 files
 set n-aliases.conf n-env.conf n-history.conf n-list.conf n-panelize.conf n-cd.conf n-functions.conf n-kill.conf n-options.conf
 
-for i; do
-    if ! test -f "$CONFIG_DIR/$i"; then
-        cp "$REPO_DIR/.config/znt/$i" "$CONFIG_DIR"
-    fi
-done
+# Check for random 2 files if they exist
+# This will shift 0 - 7 elements
+shift $(( RANDOM % 8 ))
+if ! test -f "$CONFIG_DIR/$1" || ! test -f "$CONFIG_DIR/$2"; then
+    # Something changed - examine every file
+    set n-aliases.conf n-env.conf n-history.conf n-list.conf n-panelize.conf n-cd.conf n-functions.conf n-kill.conf n-options.conf
+    for i; do
+        if ! test -f "$CONFIG_DIR/$i"; then
+            cp "$REPO_DIR/.config/znt/$i" "$CONFIG_DIR"
+        fi
+    done
+fi
 
 # Don't leave positional parameters being set
 set --
